fs/ramfs: start writing ramfs
