+++ moose/include/fs/ramfs.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <fs/vfs.h>
+
+int ramfs_mount(struct superblock *sb);
